Make the round counter a 64-bit unsigned integer. This lets us do a better -c0 (never stop sending packets), making it virtually impossible to run out of bits in the loop counter. Implementing this was suggested by Fyodor in http://seclists.org/nmap-dev/2012/q3/846

+++ nping/NpingOps.h
@@ -207,7 +207,7 @@ class NpingOps {
     bool show_eth;
 
     /* Operation and Performance */
-    u32 rounds;               /* No of times a host is targeted        */
+    u64 rounds;               /* No of times a host is targeted        */
     bool rounds_set;
     int sendpref;             /* Sending preference: eth or raw ip     */
     long host_timeout;        /* Timeout for host replies              */
@@ -351,8 +351,8 @@ class NpingOps {
     long getDelay();
     bool issetDelay();
 
-    int setRounds(u32 val);
-    u32 getRounds();
+    int setRounds(u64 val);
+    u64 getRounds();
     bool issetRounds();
 
     int setSendPreference(int v);
